Move addon-operation-invocation tasks to configured-addon superclass
