+++ app/models/configured-addon.ts
@@ -1,10 +1,33 @@
 import Model, { AsyncBelongsTo, attr, belongsTo } from '@ember-data/model';
+import { waitFor } from '@ember/test-waiters';
+import { task } from 'ember-concurrency';
 
+import { ItemType } from './addon-operation-invocation';
 import ResourceReferenceModel from './resource-reference';
 import UserReferenceModel from './user-reference';
 
+export enum ConnectedCapabilities {
+    Access = 'ACCESS',
+    Update = 'UPDATE',
+}
+
+export enum ConnectedOperationNames {
+    DownloadAsZip = 'download_as_zip',
+    CopyInto = 'copy_into',
+    HasRevisions = 'has_revisions',
+    ListRootItems = 'list_root_items',
+    ListChildItems = 'list_child_items',
+    GetItemInfo = 'get_item_info',
+}
+
+export interface OperationKwargs {
+    itemId?: string;
+    itemType?: ItemType;
+    pageCursor?: string;
+}
 export interface ConfiguredAddonEditableAttrs {
     displayName: string;
+    rootFolder: string;
 }
 
 export default class ConfiguredAddonModel extends Model {
@@ -15,10 +38,39 @@ export default class ConfiguredAddonModel extends Model {
     @attr('string') iconUrl!: string;
     @attr('string') authorizedResourceUri!: string;
 
+    @attr('array') connectedCapabilities!: ConnectedCapabilities[];
+    @attr('array') connectedOperationNames!: ConnectedOperationNames[];
+    @attr('fixstring') rootFolder!: string;
+
 
     @belongsTo('user-reference', { inverse: null })
     accountOwner!: AsyncBelongsTo<UserReferenceModel> & UserReferenceModel;
 
     @belongsTo('resource-reference', { inverse: 'configuredStorageAddons' })
     authorizedResource!: AsyncBelongsTo<ResourceReferenceModel> & ResourceReferenceModel;
+
+    @task
+    @waitFor
+    async getFolderItems(this: ConfiguredAddonModel, kwargs?: OperationKwargs) {
+        const operationKwargs = kwargs || {};
+        const operationName = operationKwargs.itemId ? ConnectedOperationNames.ListChildItems :
+            ConnectedOperationNames.ListRootItems;
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName,
+            operationKwargs,
+            thruAddon: this,
+        });
+        return await newInvocation.save();
+    }
+
+    @task
+    @waitFor
+    async getItemInfo(this: ConfiguredAddonModel, itemId: string) {
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName: ConnectedOperationNames.GetItemInfo,
+            operationKwargs: { itemId },
+            thruAddon: this,
+        });
+        return await newInvocation.save();
+    }
 }
