+++ app/models/configured-citation-addon.ts
@@ -1,6 +1,5 @@
-import { AsyncBelongsTo, attr, belongsTo } from '@ember-data/model';
+import { AsyncBelongsTo, belongsTo } from '@ember-data/model';
 
-import { ConnectedCapabilities } from 'ember-osf-web/models/configured-storage-addon';
 import AuthorizedCitationAccountModel from './authorized-citation-account';
 import ExternalCitationServiceModel from './external-citation-service';
 import ConfiguredAddonModel from './configured-addon';
@@ -11,8 +10,6 @@ export default class ConfiguredCitationAddonModel extends ConfiguredAddonModel {
 
     @belongsTo('authorized-citation-account')
     baseAccount!: AsyncBelongsTo<AuthorizedCitationAccountModel> & AuthorizedCitationAccountModel;
-
-    @attr('array') connectedCapabilities!: ConnectedCapabilities[];
 
     get externalServiceId() {
         return (this as ConfiguredCitationAddonModel).belongsTo('externalCitationService').id();
