+++ lib/osf-components/addon/components/addons-service/configured-addon-edit/template.hbs
@@ -21,7 +21,7 @@
         {{/if}}
     </div>
     <AddonsService::FileManager
-        @configuredStorageAddon={{@configuredStorageAddon}}
+        @configuredAddon={{@configuredAddon}}
         @defaultKwargs={{this.defaultKwargs}}
         @startingFolderId={{this.selectedFolder}}
         as |fileManager|
